pyproject.toml
[project]
name = "worldorder"
version = "0.1.0"
description = "AI-Powered Civilization & Geopolitical Simulation Engine"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "h4n0m4n"}]
keywords = [
    "ai", "simulation", "geopolitics", "agents", "llm",
    "civilization", "strategy", "world-leaders", "multi-agent",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment :: Simulation",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.0",
    "textual>=0.50",
    "click>=8.1",
    "chromadb>=0.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]
web = ["fastapi>=0.111", "uvicorn>=0.29"]

[project.scripts]
worldorder = "cli.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py311"
